Remove abstract topology class, replace with hwloc topology class
