+++ hpx/parallel/executors/thread_execution_information.hpp
@@ -9,7 +9,6 @@
 #include <hpx/config.hpp>
 #include <hpx/runtime/get_os_thread_count.hpp>
 #include <hpx/runtime/threads/policies/scheduler_mode.hpp>
-#include <hpx/runtime/threads/policies/topology.hpp>
 #include <hpx/runtime/threads/thread_executor.hpp>
 #include <hpx/runtime/threads/topology.hpp>
 #include <hpx/traits/is_launch_policy.hpp>
