+++ hpx/parallel/executors/v1/executor_information_traits.hpp
@@ -12,7 +12,7 @@
 
 #if defined(HPX_HAVE_EXECUTOR_COMPATIBILITY)
 #include <hpx/lcos/future.hpp>
-#include <hpx/runtime/threads/policies/topology.hpp>
+#include <hpx/runtime/threads/topology.hpp>
 #include <hpx/traits/detail/wrap_int.hpp>
 #include <hpx/traits/v1/is_executor.hpp>
 #include <hpx/util/always_void.hpp>
