+++ hpx/runtime/threads/detail/scheduled_thread_pool_impl.hpp
@@ -28,14 +28,13 @@
 #include <hpx/runtime/threads/thread_helpers.hpp>
 #include <hpx/runtime/threads/threadmanager.hpp>
 #include <hpx/state.hpp>
+#include <hpx/runtime/threads/topology.hpp>
 #include <hpx/throw_exception.hpp>
 #include <hpx/util/assert.hpp>
 #include <hpx/util/detail/yield_k.hpp>
 #include <hpx/util/unlock_guard.hpp>
 
 #include <boost/system/system_error.hpp>
-
-#include <hpx/runtime/threads/policies/hwloc_topology_info.hpp>
 
 #include <algorithm>
 #include <atomic>
