--- hpx/runtime/threads/policies/topology.hpp
@@ -1,11 +1,0 @@
-//  Copyright (c) 2007-2013 Hartmut Kaiser
-//
-//  Distributed under the Boost Software License, Version 1.0. (See accompanying
-//  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
-
-#if !defined(HPX_THREADS_POLICIES_TOPOLGY_NOV_25_2012_1036AM)
-#define HPX_THREADS_POLICIES_TOPOLGY_NOV_25_2012_1036AM
-
-#include <hpx/runtime/threads/policies/hwloc_topology_info.hpp>
-
-#endif
