+++ src/runtime/resource/detail/detail_partitioner.cpp
@@ -9,7 +9,7 @@
 #include <hpx/runtime/resource/partitioner.hpp>
 #include <hpx/runtime/runtime_fwd.hpp>
 #include <hpx/runtime/threads/detail/thread_pool_base.hpp>
-#include <hpx/runtime/threads/policies/topology.hpp>
+#include <hpx/runtime/threads/topology.hpp>
 #include <hpx/util/assert.hpp>
 #include <hpx/util/function.hpp>
 #include <hpx/util/static.hpp>
