+++ src/util/batch_environments/slurm_environment.cpp
@@ -8,7 +8,7 @@
 
 #include <hpx/config.hpp>
 #include <hpx/util/batch_environments/slurm_environment.hpp>
-#include <hpx/runtime/threads/policies/topology.hpp>
+#include <hpx/runtime/threads/topology.hpp>
 
 #include <hpx/util/assert.hpp>
 #include <hpx/util/safe_lexical_cast.hpp>
